+++ tests/testapp/migrations/0001_initial.py
@@ -34,6 +34,22 @@ class Migration(migrations.Migration):
             ],
         ),
         migrations.CreateModel(
+            name="CategoryLimitedLink",
+            fields=[
+                ("id", models.BigAutoField(auto_created=True, primary_key=True, serialize=False, verbose_name="ID")),
+                ("name", models.CharField(max_length=100)),
+                (
+                    "category",
+                    models.ForeignKey(
+                        limit_choices_to={"slug__startswith": "public"},
+                        on_delete=django.db.models.deletion.CASCADE,
+                        related_name="limited_links",
+                        to="testapp.category",
+                    ),
+                ),
+            ],
+        ),
+        migrations.CreateModel(
             name="Tag",
             fields=[
                 ("id", models.BigAutoField(auto_created=True, primary_key=True, serialize=False, verbose_name="ID")),